not split arithmetic expression
