+++ sql/expression/function/aggregation/window_framer.go
@@ -18,6 +18,7 @@ import (
 	"errors"
 	"io"
 
+	ast "github.com/dolthub/vitess/go/vt/sqlparser"
 	sqlerr "gopkg.in/src-d/go-errors.v1"
 
 	"github.com/dolthub/go-mysql-server/sql"
@@ -349,9 +350,9 @@ func (f *rangeFramerBase) NewFramer(interval sql.WindowInterval) (sql.WindowFram
 	case f.startCurrentRow:
 		startInclusion = f.orderBy
 	case f.startNPreceding != nil:
-		startInclusion = expression.NewMinus(f.orderBy, f.startNPreceding)
+		startInclusion = expression.NewArithmetic(f.orderBy, f.startNPreceding, ast.MinusStr)
 	case f.startNFollowing != nil:
-		startInclusion = expression.NewPlus(f.orderBy, f.startNFollowing)
+		startInclusion = expression.NewArithmetic(f.orderBy, f.startNFollowing, ast.PlusStr)
 	}
 
 	// TODO: how to validate datetime, interval pair when they aren't type comparable
@@ -364,9 +365,9 @@ func (f *rangeFramerBase) NewFramer(interval sql.WindowInterval) (sql.WindowFram
 	case f.endCurrentRow:
 		endInclusion = f.orderBy
 	case f.endNPreceding != nil:
-		endInclusion = expression.NewMinus(f.orderBy, f.endNPreceding)
+		endInclusion = expression.NewArithmetic(f.orderBy, f.endNPreceding, ast.MinusStr)
 	case f.endNFollowing != nil:
-		endInclusion = expression.NewPlus(f.orderBy, f.endNFollowing)
+		endInclusion = expression.NewArithmetic(f.orderBy, f.endNFollowing, ast.PlusStr)
 	}
 
 	// TODO: how to validate datetime, interval pair when they aren't type comparable
