+++ sql/rowexec/create_view_test.go
@@ -100,9 +100,10 @@ func TestReplaceExistingViewNative(t *testing.T) {
 	subqueryAlias := plan.NewSubqueryAlias("myview", "select i + 1 from mytable",
 		plan.NewProject(
 			[]sql.Expression{
-				expression.NewPlus(
+				expression.NewArithmetic(
 					expression.NewGetFieldWithTable(1, 1, types.Int32, "", "mytable", "i", true),
 					expression.NewLiteral(1, types.Int8),
+					"+",
 				),
 			},
 			plan.NewUnresolvedTable("mytable", ""),
